chore(MangaFire): Rename webViewHelper to webView
